requirements.txt
# Dependencias base del proyecto original
pillow
pandas
numpy
fastapi
uvicorn
python-multipart
opencv-python

# Nuevas dependencias para autenticación JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0

# Para desarrollo y testing
pytest==7.4.3
httpx==0.25.2
